+++ xrit-rx.py
@@ -27,7 +27,7 @@ keys = {}               # Decryption keys
 sck = None              # TCP socket object
 buflen = 892            # Input buffer length (1 VCDU)
 demux = None            # Demuxer class object
-ver = "1.0.1"           # xrit-rx version
+ver = "1.0.2"           # xrit-rx version
 
 
 def init():